+++ sale_google_map/static/src/views/google_map/google_map_renderer.js
@@ -196,6 +196,7 @@ export class GoogleMapRendererSales extends GoogleMapRenderer {
     get sidebarProps() {
         let props = super.sidebarProps;
         props.records = this.props.list.groups;
+        props.isGrouped = this.isGrouped;
         props.openCustomerSales = this.actionSeeSales.bind(this);
         return props;
     }